fix: post rebase fixes (iOS works as before)
