--- android/src/main/jni/reactnativeteleport.h‎
@@ -1,22 +1,0 @@
-#pragma once
-
-#include <ReactCommon/JavaTurboModule.h>
-#include <ReactCommon/TurboModule.h>
-#include <jsi/jsi.h>
-
-/**
- * Note this import and that it is not present in autogenerated header file
- * under android/build/generated/source/codegen/jni/reactnativeteleport.h
- *
- * Here we are overriding autogenerated component descriptors by prioritizing our custom headers via include path setup.
- */
-#include <react/renderer/components/reactnativeteleport/RNTPortalViewComponentDescriptor.h>
-#include <react/renderer/components/reactnativeteleport/RNTPortalHostViewComponentDescriptor.h>
-
-#include <memory>
-#include <string>
-
-namespace facebook::react {
-JSI_EXPORT
-std::shared_ptr<TurboModule> reactnativeteleport_ModuleProvider(const std::string &moduleName, const JavaTurboModule::InitParams &params);
-} // namespace facebook::react
