--- common/cpp/react/renderer/components/reactnativeteleport/RNTPortalHostViewShadowNode.cpp
@@ -1,8 +1,0 @@
-#include "RNTPortalHostViewShadowNode.h"
-
-namespace facebook::react
-{
-
-  extern const char PortalHostViewComponentName[] = "PortalHostView";
-
-} // namespace facebook::react
